refactor(commute): attendance에서 commute으로 이름변경
