+++ src/main/resources/static/css/admin-rule.css
