+++ src/main/resources/static/js/rule.js
