+++ src/main/resources/templates/attendance/admin_rule.html
@@ -120,7 +120,6 @@
                             <input type="date" id="validFrom" name="validFrom" class="date-input" required>
                             <span class="date-separator">~</span>
                             <input type="date" id="validTo" name="validTo" class="date-input">
-                            <span class="date-note">(종료일 없으면 비워두세요)</span>
                         </div>
                     </div>
                     <div class="form-actions">
